Add shim to preserve IPython.qt imports
